+++ CHANGELOG.md
@@ -5,3 +5,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [1.0.0]
 - Initial release
+
+## [1.1.0]
+- Rewrite of namespace detection
+- Also include autoload-dev namespaces
+- Add namespace cache for faster lookup